feat(web): a shared conversation, and an Agent's exits
An Agent is only useful once someone outside this console can reach it. This
gives an Agent its exits and gives a conversation an address a colleague can
open.

## /c/<id>

One link is one conversation: everyone who opens it lands in the same thread,
and that thread is the row the console already lists. The link is an address,
not a ticket — who may read it is decided by signing in and being a workspace
member, which the server already enforces on every call the page makes
(`requireWorkspaceMember` on the conversation, its timeline, sends, run
streams, cancels and approvals). A token would have been a second permission
to keep in step with the first.

The thread surface moves out of ConversationsPage (1849 → 546 lines) into
`components/conversation/ConversationThread.tsx` so two shells can mount one
implementation. `chrome="bare"` drops the admin topbar, the turn-nav rail and
the "view run" link: the rail would anchor to the window while the thread stays
measured, and the run link only rewrites the query string, so off the console
it did nothing but add a history entry.

Signing in returns to the link. `join-intent` generalises from
`/join-workspace` to any returnable path, so the SSO redirect and the 401
bounce round-trip too — not just the password form, which is the only one that
could have managed it alone.

## Exposure

A new tab lists an Agent's exits: the conversation (always open), Feishu, and
MCP as a placeholder. It mounts `FeishuConnectorPanel`, which has existed and
worked since the server side landed but was referenced by nothing.

A conversation is one exit, not two — the console and `/c/<id>` are two shells
on one thread behind one membership — so the share affordance is a row action
on a conversation, and copying a link needs no write access.

## Also

The three Feishu connector mutations refreshed the agent list but not the agent
detail, so a saved binding kept reading "not configured" with a dirty form
still offering to save. The sandbox send guard becomes shared, so the bare
shell refuses a send it cannot serve with the reason the console gives.

Verified: 14/14 views crawl with zero page errors; the anonymous arrival, the
no-access state and both themes measured. Reviewed adversarially by a fresh
reviewer, which confirmed the move was verbatim (16 hunks, all of them the
`chrome` prop or an export keyword) and found no access-control hole.
