+++ apps/web/src/App.tsx
@@ -4,8 +4,9 @@ import { AdminRouter } from "./pages/admin/AdminRouter"
 import { LoginPage } from "./pages/LoginPage"
 import { OnboardingPage } from "./pages/OnboardingPage"
 import { JoinWorkspaceLanding } from "./pages/JoinWorkspaceLanding"
-import { popPendingJoinIntent } from "./lib/join-intent"
+import { popPendingJoinIntent, stashReturnTo } from "./lib/join-intent"
 import { InviteAcceptPage } from "./pages/InviteAcceptPage"
+import { SharedConversationPage } from "./pages/SharedConversationPage"
 import { AuthProvider, useAuth } from "./lib/auth-context"
 import { ThemeProvider } from "./lib/theme-provider"
 import { useMyWorkspaces } from "./lib/api-workspaces"
@@ -46,6 +47,7 @@ function Root() {
   const { isLoading, isAuthenticated } = useAuth()
   const joinWsId = parseJoinWorkspaceId()
   const inviteToken = parseInviteToken()
+  const sharedConversationId = parseSharedConversationId()
 
   useEffect(() => {
     if (isLoading || !isAuthenticated) return
@@ -66,7 +68,15 @@ function Root() {
     return <LoadingScreen message={t("login.loading")} />
   }
   if (!isAuthenticated) {
+    // Sign-in can leave by SSO redirect as well as by the form, and only the
+    // form comes back to the same address on its own.
+    if (sharedConversationId !== null) stashReturnTo()
     return <LoginPage />
+  }
+  // A shared conversation needs a signed-in workspace member and nothing else,
+  // so it renders above the console rather than inside it.
+  if (sharedConversationId !== null) {
+    return <SharedConversationPage conversationId={sharedConversationId} />
   }
   return <AuthedRoot />
 }
@@ -76,6 +86,14 @@ function parseJoinWorkspaceId(): string | null {
   if (window.location.pathname !== "/join-workspace") return null
   const id = new URLSearchParams(window.location.search).get("id")
   return id && id.length > 0 ? id : null
+}
+
+function parseSharedConversationId(): string | null {
+  if (typeof window === "undefined") return null
+  const prefix = "/c/"
+  if (!window.location.pathname.startsWith(prefix)) return null
+  const id = window.location.pathname.slice(prefix.length).replace(/\/+$/, "")
+  return id.length > 0 ? id : null
 }
 
 function parseInviteToken(): string | null {
